Move all entities to the entities sub-directory
