+++ Controllers/AccountsController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using ParityService.Models.View;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
-using ParityService.Models;
+using ParityService.Models.Entities;
 using ParityService.Questrade;
 using ParityService.Managers;
