+++ Models/View/ServiceLinkViewModel.cs
@@ -1,5 +1,6 @@
 using System;
 using System.ComponentModel.DataAnnotations;
+using ParityService.Models.Entities;
 
 namespace ParityService.Models.View
 {
